final project/info (2)/info/info/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
mysqlclient==2.2.4
PyMySQL==1.1.0
python-dotenv==1.0.1
authlib==1.3.0
pyotp==2.9.0
qrcode==7.4.2
cryptography==42.0.5 